Replace iCloud sync code.
The previous iCloud sync code was based on an iOS 5 pre-release sample
from available in the Dev Forums. An updated sample was posted after
the manuscript and code were submitted to the publisher. The old sync
code could casuse random crashes due to a missing persistent store and
syncing was flaky. Sometimes data was synced, sometimes it was not.
Initial tests show the new syncing code to be more reliable.

The sync code in question is found in the method
mergeiCloudChanges:forContext:. The previous version transformed the
userInfo data into a format compatible with Core Data. This step is no
longer needed. Instead, the new code simply lets Core Data perform the
merge.