Queue iCloud merge notifications while creating persistent store.
When the persistent store is created the first time, there could be
a long delay as data is downloaded from iCloud. This can lead to
crashes if incoming merge notification messages are processed. The
crashes are caused by attempting to process the merge notification
messages while the persistent store is still being created.

To work around this problem, iCloud merge notification messages are
queued until the persistent store is ready. Once it is ready the
queued notifications are processed.